Add initial `forc test` support with single `#[test]` execution
This adds initial unit testing support to the `forc test` command behind
an `--unstable` flag. Currently, we simply execute the resulting
bytecode as if it were a script. This allows for running a single
`#[test]` function within a `library`, but isn't useful for much else
yet (hence the `--unstable` flag).

Currently, even though we generate an entry point for each test, we
don't yet have a way of enumerating all the available entry points, or a
way to have a script enter from particular entry points, or to map these
extra entry points back to a function's span or type declaration.

Logic for building and running tests is introduced in a dedicated
forc-test lib crate. This should allow crates that don't require the
entirety of `forc` but want to build/run tests to do so (e.g. the E2E
tests, sway-lsp).